refactor(bench): htmlReportからtrendReportモジュールを分離
sparkline・renderTrendSection・trend CSSをtrendReport.mjsに移動し、
escapeHtmlをformat.mjsに集約。生成HTMLは不変。
